Blindside sometimes not firing fix
I'm not positive, but I believe the reason for blindside occasionally failing for the enemy (or often, as I've experienced) is that it's checking to see if they have enough energy to fire off first. I'm not sure if this is by design or what. Or if this will fix it (I tested my other suggestion and I know it at least fixes the issue with the text being inaccessible and looping back to the menu).